pyproject.toml
[project]
name = "tthh-api"
version = "0.1.0"
description = ""
authors = [
    {name = "peterpk1",email = "phatsawut.phuket1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all] (>=0.116.1,<0.117.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pytest-asyncio (>=1.0.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "app"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
